// FE.Core.FW/src/apis/models/RegisteredCandidateTopikModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type RegisteredCandidateTopikModel = {
  id?: string | null;
  userId?: string;
  userProfileId?: string;
  examPurpose?: string | null;
  areaTest?: string;
  placeTest?: string;
  placeTestName?: string | null;
  isTestTOPIK?: boolean;
  examId?: string;
  testScheduleId?: string;
  knowWhere?: string | null;
  isPaid?: number | null;
  price?: number;
  registrationCode?: string | null;
  languageName?: string | null;
};